requirements.txt
py-cord
python-dotenv
parsedatetime
pytz
tiktoken
firebase-admin
gTTS
